Grid System 적용, 기타 refactoring

+++ Assets/C#/UI/SceneUI/UI_TownScene.cs
@@ -39,17 +39,17 @@ public class UI_TownScene : UI_Scene
 
         void OnClickedQuestButton(PointerEventData data)
         {
-            Get<UI_Page>((int)Pages.UI_Page_Town_Main).gameObject.SetActive(false);
+            Get<UI_Page>(Pages.UI_Page_Town_Main).gameObject.SetActive(false);
 
-            _currentPage = Get<UI_Page>((int)Pages.UI_Page_Town_Quest);
+            _currentPage = Get<UI_Page>(Pages.UI_Page_Town_Quest);
             _currentPage.gameObject.SetActive(true);
         }
 
         void OnClickedStoreButton(PointerEventData data)
         {
-            Get<UI_Page>((int)Pages.UI_Page_Town_Main).gameObject.SetActive(false);
+            Get<UI_Page>(Pages.UI_Page_Town_Main).gameObject.SetActive(false);
 
-            _currentPage = Get<UI_Page>((int)Pages.UI_Page_Town_Store);
+            _currentPage = Get<UI_Page>(Pages.UI_Page_Town_Store);
             _currentPage.gameObject.SetActive(true);
         }
 
@@ -57,7 +57,7 @@ public class UI_TownScene : UI_Scene
         {
             _currentPage.gameObject.SetActive(false);
 
-            _currentPage = Get<UI_Page>((int)Pages.UI_Page_Town_Main);
+            _currentPage = Get<UI_Page>(Pages.UI_Page_Town_Main);
             _currentPage.gameObject.SetActive(true);
         }
